refactor: usePermissionKeys and helper functions
